pyproject.toml
[project]
name = ""
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "mcp>=1.10.1",
  "openai>=1.93.0",
  "python-dotenv>=1.1.1",
  "requests>=2.32.4",
]
